Add files via upload
Matrix in total table please do not adjust the formulas. If more items are to be evaluated, the table in the files must be adjusted accordingly. New rating tables are no problem and can be added without any problems. But also here the formulas have to be adapted in the "total"-sheet.

=SUM(rater1!D$10;rater2!D$10;rater3!D$10;...)